Issue fixes
Description:

Due to ES Lint and previous commit to the dev for push notification unnecessary code was mixed to with the existing code and was causing the pipeline to fail while generating build.

Note:
As we can achieve In App Notification with the existing library(react-native-firebase),  'react-native-in-app-notification' library is removed.